requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# csAgent/config/config.py: 14
PyYAML == 6.0

# csAgent/datastore/elasticsearchDB/elasticsearchDB.py: 16
elasticsearch == 7.13.3

# csAgent/tracing/tracing.py: 17
jaeger_client == 4.5.0

# csAgent/datastore/mySQL/mySQL.py: 1
# csAgent/testing.py: 1
mysql_connector_python == 8.0.27

# csAgent/datastore/arangoDB/arangoDB.py: 20
# csAgent/tracing/tracing.py: 18
opentracing_instrumentation == 3.3.1

# csAgent/app.py: 19
# csAgent/testing.py: 2
pandas == 1.3.4

# csAgent/service/logger/logger.py: 17
pylogrus == 0.4.0

# csAgent/datastore/arangoDB/arangoDB.py: 15,16,18,19
# csAgent/models/config.py: 13
python_arango == 7.2.0

# csAgent/datastore/redisDB/redisDB.py: 14
redis == 4.0.2

# csAgent/app.py: 20
# csAgent/service/utils/utils.py: 15
requests == 2.25.1

# csAgent/service/utils/utils.py: 16
shortuuid == 1.0.1
